Move query to own service in datastore
