+++ modules/datastore/src/Drush.php
@@ -186,4 +186,32 @@ class Drush extends DrushCommands {
     }
   }
 
+  /**
+   * Delete jobstore entries related to a datastore.
+   */
+  protected function jobstorePrune($ref_uuid) {
+    
+    $jobs = [
+      [
+        "id" => substr(str_replace('__', '_', $ref_uuid), 0, -11),
+        "table" => "jobstore_filefetcher_filefetcher",
+      ],
+      [
+        "id" => md5($ref_uuid),
+        "table" => "jobstore_dkan_datastore_importer",
+      ],
+    ];
+
+    try {
+      foreach ($jobs as $job) {
+        \Drupal::database()->delete($job['table'])->condition('ref_uuid', $job['id'])->execute();
+        $this->logger('datastore')->notice("Successfully removed the {$job['table']} record for ref_uuid {$job['id']}.");
+      }
+    }
+    catch (\Exception $e) {
+      $this->logger('datastore')->error("Failed to delete the jobstore record for ref_uuid {$job['id']}.", $e->getMessage());
+    }
+  }
+
+
 }
